Move status logic into own function
